AlphaHeavy.FIX

Documentation

type Currency = String

type DayOfMonth = Int

type Decimal = DecimalRaw Integer

Arbitrary precision decimal type. As a rule programs should do decimal arithmetic with this type and only convert to other instances of DecimalRaw where required by an external interface.

Using this type is also faster because it avoids repeated conversions to and from Integer.

newtype Enumeration n a

Constructors

Enumeration a

data Exchange

Constructors

Exchange_NASDAQ

Exchange_NYSE

Exchange_SMART

Exchange_OTHER String

data FIXException

Constructors

DoNotSend String

FieldConvertError String

FieldNotFound !Int String

IncorrectDataFormat !Int String

IncorrectTagValue !Int String

RejectLogon String

RequiredTagMissing !Int String

UnsupportedMessageType String

newtype Field n a

Constructors

Field a

class FieldTag a where

Associated Types

type FieldTagRep a :: *

Methods

toFieldTagRep :: a -> FieldTagRep a

fromFieldTagRep :: FieldTagRep a -> Maybe a

newtype Group n a

Constructors

Group [a]

type MarketLocalTime = Int

newtype Message n dir a

Constructors

Message a

data MessageDirection

Constructors

RequestDirection

ResponseDirection

type MonthYear = Int

newtype Price

Constructors

Price Decimal

newtype Quantity

Constructors

Quantity Decimal

type UTCDate = Int

type UTCTime = Int

type UTCTimeStamp = UTCTime